26.06.2018 14:22

+++ src/com/epam/creatures/factory/CreatureFactory.java
@@ -4,14 +4,14 @@ import com.epam.creatures.entity.Creature;
 
 public class CreatureFactory {
 
-    public Creature createCreature(long creatureId, String creatureName, int limbQuantity, int headQuantity, int eyeQuantity,
-                                          Creature.Gender creatureGender, String description, Double creatureRating, Integer creatorId){
+    public Creature createCreature(Integer creatureId, String creatureName, int limbQuantity, int headQuantity, int eyeQuantity,
+                                   Creature.Gender creatureGender, String description, Double creatureRating, Integer creatorId){
 
         return new Creature(creatureId,creatureName,limbQuantity,headQuantity,eyeQuantity,creatureGender,description,creatureRating, creatorId);
     }
 
-    public Creature createCreature(long creatureId, String creatureName, int limbQuantity, int headQuantity, int eyeQuantity,
-                                          Creature.Gender creatureGender, Integer creatorId){
+    public Creature createCreature(Integer creatureId, String creatureName, int limbQuantity, int headQuantity, int eyeQuantity,
+                                   Creature.Gender creatureGender, Integer creatorId){
 
         return new Creature(creatureId,creatureName,limbQuantity,headQuantity,eyeQuantity,creatureGender, creatorId);
     }
